updated block page functionality
Updated block page functionality so that it will set a time when you go to a page that has a budget, and then blocks the page if you stay on that page longer than the budget allows.

Also fixed the annoying issue where closing the last tab wouldn't store that last tabs data in storage.  Had to add window.onunload and window.beforeunload functions in chrome.scripting function to be able to listen to when the last tab is closed, add the tab data to storage, and then retrieve and record the next time chrome is opened.

Also made it so it stops recording tab data when the computer goes idle.